pyproject.toml
[tool.poetry]
name = "bio_short_feature"
version = "0.1.0"
description = ""
authors = ["AshDavid12 <gaashdavid@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11.9"
pytube = "^15.0.0"
pydantic = "^2.8.2"
python-dotenv = "^1.0.1"
fastapi = "^0.111.1"
requests = "^2.32.3"
httpx = "^0.27.0"
uvicorn = "^0.30.1"
asyncio = "^3.4.3"
confluent-kafka = "^2.5.0"
aiokafka = "^0.11.0"
streamlit = "1.25.0"
openai = "^1.37.0"
langchain = "^0.2.11"
langchain-openai = "^0.1.17"
langchain-core = "^0.2.23"
cohere = "^5.6.2"
langchain-cohere = "^0.1.9"
langserve = "^0.2.2"
sse-starlette = "^2.1.2"
langchain-community = "^0.2.10"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
